requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
gTTS
yt-dlp
qrcode
Pillow
selenium
